todo_tasks/simple_todo/requirements.txt
django==3.1.5
djangorestframework
